🚚 zb: Move socket::Split into a separate mod
